04-microservice/test/notes/requirements.txt
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
Deprecated==1.2.13
essential-generators==1.0
Flask==2.2.2
googleapis-common-protos==1.56.2
grpcio==1.49.1
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
opentelemetry-api==1.13.0
opentelemetry-exporter-jaeger==1.13.0
opentelemetry-exporter-jaeger-proto-grpc==1.13.0
opentelemetry-exporter-jaeger-thrift==1.13.0
opentelemetry-instrumentation==0.34b0
opentelemetry-instrumentation-flask==0.34b0
opentelemetry-instrumentation-pymongo==0.34b0
opentelemetry-instrumentation-wsgi==0.34b0
opentelemetry-sdk==1.13.0
opentelemetry-semantic-conventions==0.34b0
opentelemetry-util-http==0.34b0
protobuf==3.20.3
pymongo==4.2.0
python-dateutil==2.8.2
requests==2.28.1
six==1.16.0
thrift==0.16.0
typing_extensions==4.4.0
urllib3==1.26.12
Werkzeug==2.2.2
wrapt==1.14.1
